fix(presence): compute payload-size guard in uint64 for 32-bit builds
`len(payload) > math.MaxUint32` fails to compile on 32-bit targets
(linux/386, windows/386, 32-bit arm): int is 32-bit there and the untyped
constant 4294967295 overflows it. Do the comparison in uint64 space.

+++ internal/presence/client.go
@@ -294,7 +294,7 @@ func writeJSONFrame(writer io.Writer, opcode uint32, value any) error {
 	if err != nil {
 		return fmt.Errorf("encode JSON: %w", err)
 	}
-	if len(payload) > math.MaxUint32 {
+	if uint64(len(payload)) > math.MaxUint32 {
 		return fmt.Errorf("JSON payload is too large: %d bytes", len(payload))
 	}
